Reorganize OSLCommon folder structure
