+++ OSLCommon/Authorization/Account.cs
@@ -13,45 +13,6 @@ using System.Xml.Serialization;
 
 namespace OSLCommon.Authorization
 {
-    public class AccountDisableReason
-    {
-        public string Message = "";
-        public long Timestamp = 0;
-    }
-    public class AccountTokenDetailsResponse
-    {
-        public string Username { get; set; }
-        public bool Enabled { get; set; }
-        public long CreatedTimestamp { get; set; }
-        public long LastUsed { get; set; }
-        public string UserAgent { get; set; }
-        public string Host { get; set; }
-        public string Hash { get; set; }
-    }
-    public class AccountDetailsResponse
-    {
-        public AccountDetailsResponse()
-        {
-            Username = "";
-            Enabled = true;
-            Permissions = Array.Empty<AccountPermission>();
-            DisableReasons = Array.Empty<AccountDisableReason>();
-            Licenses = Array.Empty<string>();
-            Groups = Array.Empty<string>();
-            FirstSeenTimestamp = 0;
-            LastSeenTimestamp = 0;
-            Tokens = Array.Empty<AccountTokenDetailsResponse>();
-        }
-        public string Username { get; set; }
-        public bool Enabled { get; set; }
-        public AccountPermission[] Permissions { get; set; }
-        public AccountDisableReason[] DisableReasons { get; set; }
-        public string[] Licenses { get; set; }
-        public string[] Groups { get; set; }
-        public long FirstSeenTimestamp { get; set; }
-        public long LastSeenTimestamp { get; set; }
-        public AccountTokenDetailsResponse[] Tokens { get; set; }
-    }
     public class Account
     {
         #region Constructors
