+++ OSLCommon/Authorization/AccountDisableReason.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace OSLCommon.Authorization
+{
+    public class AccountDisableReason
+    {
+        public string Message = "";
+        public long Timestamp = 0;
+    }
+}
