+++ OSLCommon/Authorization/AccountManager.cs
@@ -1,5 +1,6 @@
 ﻿using Google.Cloud.Firestore.V1;
 using kate.shared.Helpers;
+using OSLCommon.AuthProviders;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
