+++ OSLCommon/Endpoint.cs
@@ -1,5 +1,6 @@
 ﻿using kate.shared.Helpers;
 using OSLCommon.Authorization;
+
 using System;
 using System.Collections.Generic;
 using System.Net;
