+++ OSLCommon/EndpointShared.cs
@@ -1,5 +1,6 @@
 ﻿using OSLCommon.Authorization;
 using OSLCommon.AutoUpdater;
+
 using System;
 using System.Collections.Generic;
 using System.Text;
