+++ OpenSoftwareLauncher.Server/Properties/VersionAutoIncrement.cs
@@ -1,6 +1,6 @@
 ﻿
 using System.Reflection;
 
-[assembly: AssemblyVersion("1.0.109.225")]
-[assembly: AssemblyFileVersion("1.0.109.225")]
+[assembly: AssemblyVersion("1.0.112.271")]
+[assembly: AssemblyFileVersion("1.0.112.271")]
